+++ pde/laplace.py
@@ -16,7 +16,7 @@ __all__ = ['solve']
 
 _METHODS = ['ic']
 
-def solve(domain, conds, method='ic'):
+def solve(axis, conds, method='ic'):
     """
     Methods
     -------
@@ -24,39 +24,39 @@ def solve(domain, conds, method='ic'):
 
     Parameters
     ----------
-    domain : array_like
-        [xn, xf, yn, yf], 'xn' and 'yn' are the number of partitions at
-        axis 'x' and 'y', 'xf' and 'yf' are the final positions; [int,
-        float, int, float].
+    axis : array_like
+        Axis 'x' and 'y'; [x, y], each element should be an array_like.
     conds : array_like
         Boundary conditions; [bound_x0, bound_xf, bound_y0, bound_yf],
-        each element should be a scalar or an array_like of size 'xn+1'
-        for 'cond_x' and size 'yn+1' for 'cond_y'.
+        each element should be a scalar or an array_like of size of 'x'
+        for 'cond_y's and size of 'y' for 'cond_x's.
     method : string | optional
         Finite-difference method.
 
     Returns
     -------
     u : ndarray
-        A matrix of size (xn+1)*(yn+1); u[x, y].
+        A 2-D ndarray; u[x, y].
     """
     base.check_method(method, _METHODS)
 
-    u = steady.set_u(*domain[::2], *conds)
-    consts = _cal_constants(*domain)
+    u = steady.set_u(*axis, *conds)
+    consts = _cal_constants(*axis)
 
-    _implicit(u, *domain[::2], *consts)
+    _implicit(u, *axis, *consts)
 
     return u
 
-def _implicit(u, xn, yn, 𝛂, β):
+def _implicit(u, x, y, 𝛂, β):
     """Métodos de diferenças finitas implícitos."""
-    mat = _set_mat(𝛂, β, xn, yn)
+    xn, yn = x.size, y.size
+
+    mat = _set_mat(𝛂, β, xn-1, yn-1)
     vec = _set_vec(𝛂, β, u)
 
     x = linalg.solve(mat, vec)
 
-    u[1:-1, 1:-1] = np.reshape(x, (xn-1, yn-1), 'F')
+    u[1:-1, 1:-1] = np.reshape(x, (xn-2, yn-2), 'F')
 
 def _set_mat(𝛂, β, xn, yn):
     """Monta a matriz do sistema em '_implicit()'."""
@@ -82,9 +82,9 @@ def _set_vec(𝛂, β, u):
 
     return np.reshape(vec, np.size(vec), 'F')
 
-def _cal_constants(xn, xf, yn, yf):
+def _cal_constants(x, y):
     """Calcula as constantes '𝛂' e 'β'."""
-    𝛂 = (xf/xn) ** 2
-    β = (yf/yn) ** 2
+    𝛂 = (x[-1] / (x.size-1)) ** 2
+    β = (y[-1] / (y.size-1)) ** 2
 
     return (𝛂, β)